tere/pyproject.toml
[project]
name = "tere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
]
